Fix one failing diff test
This test previously directly called the diff method which we've deleted. Rewriting the test to call our new ConfigurationDiffer. The test required `translate_yaml_input` so we have made it public. A future improvement as part of this piece of work is to make that method public and tested anyway, so it is fine to do this now.