Move Vcloud::QueryRunner under Vcloud::Core
Move Vcloud::QueryRunner under Vcloud::Core to comply with breaking
changes introduced in vcloud-core version 0.2.0.

Also bumps the dependency on vcloud-core in the Gemspec so that we use
version 0.2.0, and bumps vcloud-net_launcher to version 0.1.0.

Prefer a minor version bump for vcloud-net_launcher (as opposed to a
patch version bump) to make it clearer that the changes in vcloud-core
broke backwards compatibility and that it's preferable to upgrade.