Add unit tests for NetLaunch
This is the only project that didn't have unit test. Mostly because it's the
simplest - has no real logic of it's own, just uses Vcloud::Core. But it
would still benefit from having one. Mostly ripped off from Vcloud::Launcher
with the exceptions that:

- It calls describe on the object to be tested rather than namespacing it
  within a module as we plan to do for the other projects.
- Uses `let!` to separate variables from logic/expectations.
- It doesn't continue on errors like Vcloud::Launcher can.

It has the added benefit that:

- `rake spec` no longer executes all of the tests because the `Filelist[]`
  doesn't return any matches.
- We can make the testing section of the README look more like the other
  projects.