Fix 'expect() without .to' assertation error
We've seen this common error in other tests. When `expect ==` is used
without the '.to', the test passes regardless.

Eg

    expect(false) == true    # passes

This pitfall can be avoided by always using a matcher, eg:

    expect(false) eq(true)   # gives syntax error

This commit fixes all assertations to use:

    expect(thing).to eq(value)

It has highlighted some broken tests. These will be fixed up in the
following commit.