Document public methods with Rdoc
Document the public API methods using Rdoc.

Note I have not documented `NetLaunch#new` as there was nothing I could
add that would not repeat the documentation for `NetLaunch#run` and I
think it's simple enough so as to be self-explanatory.

Also, YARD (though not Rdoc) will automatically document it as:

> A new instance of NetLaunch.