Allow anchors to be created under a dedicated key
It is sometimes useful to extract common config into YAML anchors. Currently
the config schema is very strict, and fails if any additional keys are present
in the config file. This change would allow an open-ended `anchors` key to
exist in the config file, under which any number of anchors can be created.

Example:

```
---
anchors:
  common: &common
    edge_gateway: "nft0000000"
    vdc_name: "my-vdc"
    fence_mode: 'natRouted'
    netmask: '255.255.255.0'

org_vdc_networks:
  - name: 'app'
    description: 'app network'
    gateway: '10.0.0.1'
    ip_ranges:
      - start_address: '10.0.0.2'
      - end_address: '10.0.0.254'
    <<: *common
```